[flang][OpenMP] Generate implicit mapping for reduction variable. (#15)
* Generate implicit mapping for reduction variable.

The result of the reduction was not being copied back to reduction variable on the host. To sidestep this problem, an explicit map(tofrom: var) has to be added in the code to get the result back.

This patch tries to solve this problem by generating an implicit tofrom mapping for reduction variable. All variables used in the target region (and without an explicit mapping) gets an implcit mapping in getTargetOP. We switch on TO,FROM bits for reduction varibles on this mapping.

* Move code to get the reduction variables into separate function.