[flang][OpenMP] Add initial pass to map `do concurrent` to OMP
Adds a new pass to map `do concurrent` loops to OpenMP. So far the pass
only maps to the host, i.e. `parallel do`. Later, it will be extended to
map to the device/target as well.

This diff also adds the pass the code-gen pipeline flang. To invoke the
new pass from flang, use `-fdo-concurrent-parallel=<none|host|device>`.
The option is only active when OpenMP is enabled throught `-fopenmp`.