[flang][OpenMP] Privatize locally destroyed values in `do concurent` (#112)
Collects values that are local to a loop: "loop-local values". A loop-local
value is one that is used exclusively inside the loop but allocated outside
of it. This usually corresponds to temporary values that are used inside the
loop body for initialzing other variables for example. For a "loop-local" 
value within a loop's scope, localizes that value within the scope of the 
parallel region the loop maps to.