[Flang] Disable Global Filtering pass by default (#236)
Global Filtering pass was created to fix linking issues with the old clang driver.
This pass has not been upstreamed yet and can be switched off for flang-new
when used with the new clang driver. However, the user can still turn the pass
on by setting the `-foffload-global-filtering` flag.

This commit marks the first step towards completely removing the global filtering pass
from the ATD branch and minimizing the delta between ATD and upstream branches.